missing files

+++ albinaria-dataset/proc/scripts/build_consensus_loci.sh
@@ -1,24 +1,37 @@
-kindex=$1
+#!/bin/sh
+
+PREFIX=$1
+kindex=$2
+RES_DIR=$3
+LOG_DIR=$4
+OUT_DIR=$5
+
+[ -z $OUT_DIR ] && exit
 
 RAXML_BIN=raxmlHPC-SSE3
-resdir=res/$kindex
-logdir=log/consensus/$kindex
-outfile=consensus.$kindex.trees
+resdir=$RES_DIR/$kindex
+logdir=$LOG_DIR/consensus/$kindex
+outfile=$OUT_DIR/consensus.$kindex.trees
+tmpfile=/tmp/tmp.lociconsensus.$kindex
+
+CONSENSUS_TYPE=MRE
+RAXML_RESULT_PREFIX=RAxML_MajorityRuleExtendedConsensusTree
 
 mkdir -p $logdir
 rm -f $outfile
 
-kstart=$((kindex * 1000))
-kend=$(((kindex+1) * 1000 - 1))
-for i in `seq $kstart $kend`; do
-  tfile=`ls $resdir/*.locus.$i.trees 2> /dev/null`
-  if [ ! -z $tfile -a -f $tfile ]; then
-    execname=strict.$i
-    $RAXML_BIN -m GTRGAMMA -J STRICT -z $tfile -n $execname 2>&1 > /dev/null
-    tree=`cat RAxML_StrictConsensusTree.$execname`
-    echo $i $tree >> $outfile
-    mv RAxML_info.$execname $logdir
-    mv RAxML_StrictConsensusTree.$execname $logdir
-    echo $i
-  fi
+for tfile in $resdir/$PREFIX.locus.*.trees; do
+  i=`echo $tfile | rev | cut -d'.' -f2 | rev`
+  echo "build consensus for $tfile : $i"
+  execname=locusmre.$i
+  $RAXML_BIN -m GTRGAMMA -J $CONSENSUS_TYPE -z $tfile -n $execname 2>&1
+  tree=`cat $RAXML_RESULT_PREFIX.$execname`
+  echo $i $tree >> $outfile
+  mv RAxML_info.$execname $logdir
+  mv $RAXML_RESULT_PREFIX.$execname $resdir
+  echo "done $i"
 done
+
+mv $outfile $tmpfile
+sort -n $tmpfile > $outfile
+rm $tmpfile
